kv_cache mixin
