+++ olive/model/handler/mixin/__init__.py
@@ -7,6 +7,7 @@ from olive.model.handler.mixin.dummy_inputs import DummyInputsMixin
 from olive.model.handler.mixin.hf_config import HfConfigMixin
 from olive.model.handler.mixin.io_config import IoConfigMixin
 from olive.model.handler.mixin.json import JsonMixin
+from olive.model.handler.mixin.kv_cache import PytorchKvCacheMixin
 from olive.model.handler.mixin.onnx_ep import OnnxEpValidateMixin
 from olive.model.handler.mixin.onnx_graph import OnnxGraphMixin
 from olive.model.handler.mixin.resource import ResourceMixin
@@ -19,5 +20,6 @@ __all__ = [
     "JsonMixin",
     "OnnxEpValidateMixin",
     "OnnxGraphMixin",
+    "PytorchKvCacheMixin",
     "ResourceMixin",
 ]
